Moving route poller to lib function, deleting API
