+++ CHANGES.md
@@ -1,7 +1,6 @@
 # List of changes to be made in the 2.0 release
 - change drawer search to form, throw redirect on server, consider enhancing
 - change strike to server load, load everything in parent load, reuse in component, remove API
-- change route to library function, call from wherever is needed, remove API
 - consider moving all pollers to separate library directory
 - investigate possibility of using enhanced forms with method="GET" with search queries
