+++ src/routes/metro/[stop]/map/+page.server.ts
@@ -1,4 +1,5 @@
 import type { stopDB } from '$lib/stopDB.js';
+import type { LatLngExpression } from 'leaflet';
 
 const STOP_NUM = 50;
 
@@ -6,6 +7,8 @@ export async function load({ locals, parent }) {
     const parentData = await parent();
 
     const { stops } = locals;
+
+    const coords = [parentData.db.coordinates[1], parentData.db.coordinates[0]] as LatLngExpression;    //Need to swap coords, mongo wants [lon, lat]; leaflet wants [lat, lon]
 
     const query = { coordinates: { $nearSphere: { $geometry: { type: 'point', coordinates: parentData.db.coordinates } } }, code: { $ne: parentData.code } }
     const projection = { _id: 0, city: 0 }
@@ -22,6 +25,7 @@ export async function load({ locals, parent }) {
     })) satisfies stopDB[];
 
     return {
-        near
+        near,
+        coords
     };
 }
