+++ src/routes/metro/[stop]/map/+page.svelte
@@ -1,11 +1,9 @@
 <script lang="ts">
     import 'leaflet/dist/leaflet.css';
     import { onMount } from 'svelte';
-	import { invalidate } from '$app/navigation';
     import type { PageData } from "./$types";
     import type { Marker, LatLngTuple, Map } from "leaflet";
     import type { stopDB } from "$lib/stopDB"
-	import { encodeRoute, type vehicleMap } from '$lib/vehicle';
 
     export let data: PageData;
 
@@ -13,13 +11,10 @@
     
     let mapElement: HTMLElement;
     let map: Map;
-    const markers = new Array<{droplet: Marker, vehicle: Marker, code: number}>;
     let loaded = false;
-    let routes = new Array<vehicleMap>;
     
     const pinColour = '#1b8ae8';
     const otherPinColour = '#909090';
-    const REFRESH_TIME = 1000;
 
     onMount(async () => {
         const L = await import('leaflet');  //Leaflet has to be imported here, it needs window to be defined
@@ -42,48 +37,7 @@
             L.marker(stop.coordinates as LatLngTuple, {icon: otherPinIcon}).addTo(map).bindPopup(getPopup(stop));    
         }
 
-        const passages = await data.vehicles.promise;   //Wait for vehicle data to arrive
-        routes = passages;
         loaded = true;
-
-        //Place vehicle icons
-        for(const pass of passages){
-            const { busIcon, tramIcon, dropletIcon } = getVehicleIcons(L, pass.colour);
-            for(const vehicle of pass.vehicles){
-                const vehicleIcon = vehicle.vehicleType === 'Tram' ? tramIcon : busIcon;
-                const popup = `<a href="/route/${encodeRoute(pass.routeID)}"><div>Linea ${pass.route}</a><br><a href="/vehicle/${vehicle.id}">${vehicle.vehicleType} ${vehicle.id}</div></a>`;
-
-                const dropletMark = L.marker([vehicle.lat, vehicle.lon], {icon: dropletIcon, zIndexOffset: 100, alt: vehicle.vehicleType + ' ' + vehicle.id, rotationAngle: vehicle.direction}).addTo(map).bindPopup(popup);
-
-                const vehicleMark = L.marker([vehicle.lat, vehicle.lon], {icon: vehicleIcon, zIndexOffset: 101}).addTo(map).bindPopup(popup);
-                
-                markers.push({
-                    droplet: dropletMark,
-                    vehicle: vehicleMark,
-                    code: vehicle.id
-                });
-            }
-        }
-
-        //Refresh vehicles positions
-        setInterval(async() => {
-            invalidate('stop_lines');
-            const vehicles = await data.vehicles.promise;
-            routes = vehicles;
-            
-            for(const route of vehicles){
-                for(const vehicle of route.vehicles){
-                    for(const marker of markers){
-                        if(marker.code === vehicle.id){
-                            marker.droplet.setLatLng([vehicle.lat, vehicle.lon]);
-                            marker.vehicle.setLatLng([vehicle.lat, vehicle.lon]);
-                            marker.droplet.setRotationAngle(vehicle.direction);                        
-                        }
-                    }
-                }
-            }
-        }, REFRESH_TIME);
-
     });
 
     //Return the appropriate popup link for a stop, depending on whether it's a regular stop, metro station or train station
